Fix reading from cool for neighbouring runs (TrigConfStorage-03-00-88)
	* Fix reading from cool for neighbouring runs
	* TrigConfStorage-03-00-88

2015-02-19  Werner Wiedenmann  <Werner.Wiedenmann@cern.ch>

	* TrigConfStorage-03-00-87
	* src/ReplicaSorter.cxx: remove printout statements (JIRA ATR-10212)
	* src/ReplicaSorter.cxx: use gethostname (JIRA ATR-10222)

2015-02-19 Joerg Stelzer
	* Update of COOL fix
	* TrigConfStorage-03-00-86

2015-02-17 Ignacio Aracena
	* src/L1TopoMenuLoader.cxx
	* Change from int to long : TMT_VERSION, TP_VALUE
	* ATR-10114
	* TrigConfStorage-03-00-85

2015-02-12 Elisabetta Pianori
...
(Long ChangeLog diff - truncated)

+++ Trigger/TrigConfiguration/TrigConfStorage/src/L1TopoMenuLoader.cxx
@@ -72,7 +72,7 @@ bool TrigConf::L1TopoMenuLoader::loadTopoAlgos( TXC::L1TopoMenu& tcaTarget) {
 
       coral::AttributeList attList0;
       attList0.extend<int>( "TMT.TMT_TRIGGER_MENU_ID" );
-      attList0.extend<int>( "TMT.TMT_VERSION" );
+      attList0.extend<long>( "TMT.TMT_VERSION" );
       attList0.extend<string>( "TTM.TTM_NAME" );
       attList0.extend<int>( "TTM.TTM_VERSION" );
       attList0.extend<int>( "TTM.TTM_CTPLINK_ID" );
@@ -187,7 +187,7 @@ bool TrigConf::L1TopoMenuLoader::loadAlgRegister( TXC::L1TopoConfigAlg& tcaTarge
 
       coral::AttributeList attList0;
       attList0.extend<string>( "TP.TP_NAME" );
-      attList0.extend<int>( "TP.TP_VALUE" );
+      attList0.extend<long>( "TP.TP_VALUE" );
       attList0.extend<int>( "TP.TP_POSITION" );
       attList0.extend<int>( "TP.TP_SELECTION" );
       query0->defineOutput(attList0);     
@@ -198,7 +198,7 @@ bool TrigConf::L1TopoMenuLoader::loadAlgRegister( TXC::L1TopoConfigAlg& tcaTarge
          const coral::AttributeList& row0 = cursor0.currentRow();
          //std::cout << "loadAlgRegister my row " << row0 << std::endl;
          string tp_name = row0["TP.TP_NAME"].data<string>();
-         int tp_value = row0["TP.TP_VALUE"].data<int>();
+         long tp_value = row0["TP.TP_VALUE"].data<long>();
          int tp_pos = row0["TP.TP_POSITION"].data<int>();
          int tp_sel = row0["TP.TP_SELECTION"].data<int>();
          tcaTarget.addParameter(tp_name,to_string(tp_value),tp_pos, tp_sel);
